feat: added Roles

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
@@ -12,7 +13,9 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::all();
+        $users = User::orderBy('name')->get();
+
+        return response()->json($users, 200);
     }
 
     /**
@@ -20,7 +23,36 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        return User::create($request->all());
+
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|string|email|unique:users',
+            'name' => 'required|string|max:100',
+            'phone_number' => 'nullable|digits:10',
+            'identification_card' => 'required|string|max:11',
+            'birth_date' => 'required|date|before_or_equal:' . now()->subYears(18)->format('Y-m-d'),
+            'city_code' => 'required|integer|max:999999',
+            'role_id' => 'required|exists:roles,id',
+            'password' => 'required|string|min:8',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+
+        $user = User::create([
+            'email' => $request->email,
+            'name' => $request->name,
+            'phone_number' => $request->phone_number,
+            'identification_card' => $request->identification_card,
+            'birth_date' => $request->birth_date,
+            'city_code' => $request->city_code,
+            'password' => bcrypt($request->password),
+        ]);
+
+        $user->roles()->attach($request->role_id);
+
+
+        return response()->json($user, 201);
     }
 
     /**
@@ -28,7 +60,15 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
-        return User::find($id);
+
+        $user = User::find($id);
+
+        if (!$user) {
+            return response()->json(['error' => 'Usuario no encontrado'], 404);
+        }
+
+
+        return response()->json($user, 200);
     }
 
 
@@ -38,17 +78,52 @@ class UserController extends Controller
     public function update(Request $request, string $id)
     {
         $user = User::findOrFail($id);
-        $user->update($request->all());
-        return $user;
+
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|string|email|unique:users,email,' . $user->id,
+            'name' => 'required|string|max:100',
+            'phone_number' => 'nullable|digits:10',
+            'identification_card' => 'required|string|max:11',
+            'birth_date' => 'required|date|before_or_equal:' . now()->subYears(18)->format('Y-m-d'),
+            'city_code' => 'required|integer|max:999999',
+            'role_id' => 'required|exists:roles,id',
+            'password' => 'sometimes|required|string|min:8',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+
+        $user->update([
+            'email' => $request->email,
+            'name' => $request->name,
+            'phone_number' => $request->phone_number,
+            'identification_card' => $request->identification_card,
+            'birth_date' => $request->birth_date,
+            'city_code' => $request->city_code,
+            'password' => isset($request->password) ? bcrypt($request->password) : $user->password,
+        ]);
+
+        $user->roles()->sync([$request->role_id]);
+
+        return response()->json($user, 200);
     }
+
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
     {
-        $user = User::findOrFail($id);
+
+        $user = User::find($id);
+
+        if (!$user) {
+            return response()->json(['error' => 'Usuario no encontrado.'], 404);
+        }
+
         $user->delete();
-        return 204;
+
+        return response()->json(null, 204);
     }
 }
